mori/requirements.txt
torch==2.3.1+cpu
torchvision==0.18.1+cpu
cryptography==44.0.2
deep-translator==1.11.4
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-yasg==1.21.9
faiss-cpu==1.10.0
open_clip_torch==2.31.0
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
tqdm==4.67.1
transformers==4.49.0
psycopg2-binary
sentence-transformers